systemcentervirtualmachinemanager models update

+++ internal/services/systemcentervirtualmachinemanager/system_center_virtual_machine_manager_server_resource.go
@@ -28,7 +28,7 @@ type SystemCenterVirtualMachineManagerServerModel struct {
 	Fqdn              string            `tfschema:"fqdn"`
 	Username          string            `tfschema:"username"`
 	Password          string            `tfschema:"password"`
-	Port              int               `tfschema:"port"`
+	Port              int64             `tfschema:"port"`
 	Tags              map[string]string `tfschema:"tags"`
 }
 
@@ -142,7 +142,7 @@ func (r SystemCenterVirtualMachineManagerServerResource) Create() sdk.ResourceFu
 			}
 
 			if v := model.Port; v != 0 {
-				parameters.Properties.Port = pointer.To(int64(v))
+				parameters.Properties.Port = pointer.To(v)
 			}
 
 			if err := client.CreateOrUpdateThenPoll(ctx, id, *parameters); err != nil {
@@ -182,7 +182,7 @@ func (r SystemCenterVirtualMachineManagerServerResource) Read() sdk.ResourceFunc
 				state.CustomLocationId = pointer.From(model.ExtendedLocation.Name)
 				state.Fqdn = model.Properties.Fqdn
 				state.Password = metadata.ResourceData.Get("password").(string)
-				state.Port = int(pointer.From(model.Properties.Port))
+				state.Port = pointer.From(model.Properties.Port)
 				state.Tags = pointer.From(model.Tags)
 
 				if v := model.Properties.Credentials; v != nil {
